refactor(modelmigration): split controller model info export helpers
Break up GetControllerModelInfo so statement preparation, row
collection, and final assembly are handled
by smaller focused helpers, making the controller fact export path
easier to read without changing
behavior.
